feat: tolerate dirty json (#431)
When you use the `aws logs tail` command, it outputs non-JSON along with
the JSON log entries. For example:

```bash
2024-01-01T19:52:19 {"level":20,"time":1704138739937,"pid":8,"hostname":"169.254.43.189","source":"search/src/lib/typesense.ts","msg":"found 20 records not in index"}
```

This change allows piping this type of output and having pjl still be
able to parse the log entry. In theory, we could output the extraneous
non-JSON data to stdout, but I don't personally need that, so I didn't
implement it.

FYI, I use pino to output JSON log entries from an AWS Lambda function.

---------

Co-authored-by: Blayne Chard <blayne@chard.com>